requirements.txt
sentence-transformers
faiss-cpu
openai
langchain
tiktoken
fastapi
uvicorn
pydantic
python-multipart
python-dotenv

